Update datetime component, create composable for pay fees.
