+++ btr-web/btr-main-app/composables/usePayFeesBtr.ts
@@ -1,0 +1,15 @@
+export const addBtrPayFees = async () => {
+  const payFeesWidget = usePayFeesWidget()
+
+  const feeInfo = await payFeesWidget.getFeeInfo(
+    {
+      entityType: 'BTR',
+      filingTypeCode: 'REGSIGIN',
+      futureEffective: false,
+      priority: false,
+      waiveFees: false
+    }
+  )
+
+  return payFeesWidget.addFee(feeInfo)
+}
