+++ btr-web/btr-main-app/pages/ownerChange.vue
@@ -11,7 +11,7 @@
       <div class="ml-8 flex-auto">
         <BcrosInputsDateSelect
           :max-date="new Date()"
-          @selection="currentSIFiling.effectiveDate = dateToString($event)"
+          @selection="significantIndividualChangeDate($event)"
         />
       </div>
     </div>
@@ -40,6 +40,11 @@ const significantIndividuals = useSignificantIndividuals()
 const { currentSIFiling } = storeToRefs(significantIndividuals)
 
 const expandNewSI = ref(false)
+
+const significantIndividualChangeDate = (event) => {
+  currentSIFiling.value.effectiveDate = datetimeStringToDateString(event)
+  addBtrPayFees()
+}
 
 function addNewSI (si: SignificantIndividualI) {
   significantIndividuals.filingAddSI(si)
